backend/requirements.txt
flask
pytest
flask_restful
flask-bcrypt
PyJWT==1.7.1
Flask-Mail
flask-cors
